Remove exit 0 at the end, and sleep 30 at start
Exit 0 causes Jenkins to conclude/stop the build successfully.
Which is not the behavior I want.

There does not appear to be a reason to have the preprocessor
script sleep for 30 seconds. Unlike like a shell provisioner,
which has to wait for services/etc. to come online in a VM.